split up test file
